% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.r
\name{plot_maps}
\alias{plot_maps}
\title{{Plot standard maps}}
\usage{
plot_maps(plot_set = 3, Report, Sdreport = NULL, Xlim = NULL,
  Ylim = NULL, TmbData = NULL, spatial_list = NULL,
  Panel = "Category", FileName = paste0(getwd(), "/"),
  format = "png", category_names = NULL, covar_names = NULL,
  legend = TRUE, textmargin = NULL, ...)
}
\arguments{
\item{plot_set}{integer-vector defining plots to create
\describe{
  \item{plot_set=1}{Probability of encounter/non-encounter}
  \item{plot_set=2}{Log-expected positive catch rate}
  \item{plot_set=3}{Log-predicted density (product of encounter probability and positive catch rates)}
  \item{plot_set=4}{Log-positive catch rates (rescaled)}
  \item{plot_set=5}{Log-predicted density (rescaled)}
  \item{plot_set=6}{Spatio-temporal variation in encounter probability}
  \item{plot_set=7}{Spatio-temporal variation in log-positive catch rates}
  \item{plot_set=8}{Linear predictor for encounter probability}
  \item{plot_set=9}{Linear predictor for positive catch rates}
  \item{plot_set=10}{Coefficient of variation for predicted density (available only if \code{Data_Fn(...,Options=c('SD_site_logdensity'=1,...))}}
  \item{plot_set=11}{Covariates that are included in the model}
  \item{plot_set=12}{Total biomass across all categories (only useful in a multivariate model)}
  \item{plot_set=13}{Covariate effects on encounter probability}
  \item{plot_set=14}{Covariate effects on positive catch rates}
  \item{plot_set=15}{Individual covariate effects on encounter probability}
  \item{plot_set=16}{Individual covariate effects on positive catch rates}
}}

\item{Report}{tagged list of outputs from TMB model via \code{Obj$report()}}

\item{Sdreport}{Standard deviation outputs from TMB model via \code{sdreport(Obj)}}

\item{Xlim}{xlimits for each panel}

\item{Ylim}{ylimits for each panel}

\item{TmbData}{optional tagged list of outputs from fit$data_list or VAST::make_data}

\item{spatial_list}{required for Method == Stream_network, optional for other spatial models, tagged list of outputs from \code{make_spatial_info}}

\item{Panel}{Whether to plot years for a given category (\code{Panel="Category"}) or categories for a given year ((\code{Panel="Year"})  in each panel figure}

\item{FileName}{Directory (absolute path) and base for filenames of plots}

\item{format}{image format}

\item{category_names}{character vector specifying names for different categories (only used for R package \code{VAST})}

\item{covar_names}{character vector specifying covariate names for labeling figures}

\item{legend}{Boolean whether to plot legend or not}

\item{textmargin}{option to include y-axis text}

\item{...}{arguments passed to \code{par}}
}
\value{
Mat_xt a matrix (rows: modeled knots; column: modeled year) for plotted output of last element of \code{plot_set}
}
\description{
{plots a standard set of diagnostic maps}
}
