% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot Pearson residuals on map}
\usage{
plot_residuals(fit, Data, Network_sz_LL, category_names, FilePath)
}
\arguments{
\item{fit}{model fit from fit_model}

\item{Data}{input data in format of Data_Geostat}

\item{Network_sz_LL}{stream network info}

\item{category_names}{category names}

\item{FilePath}{path to save figures}
}
\value{
A tagged list of Pearson residuals
\describe{
  \item{Q1_xt}{Matrix of average residuals for encounter/non-encounter component by site \code{x} and year \code{t}}
  \item{Q2_xt}{Matrix of average residuals for positive-catch-rate component by site \code{x} and year \code{t}}
}
}
\description{
\code{plot_residuals} shows average Pearson residual for every knot for encounter probability and positive catch rate components
}
