% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot stream network and optional observations}
\usage{
plot_network(Network_sz_LL, Data_Geostat = NULL, byYear = FALSE,
  FilePath = NULL, FileName = "Network", arrows = FALSE,
  root = FALSE)
}
\arguments{
\item{Data_Geostat}{data frame with observations}

\item{byYear}{plot observations by year, default = FALSE}

\item{FilePath}{path to figure directory}

\item{FileName}{label for plot}

\item{arrows}{default = FALSE do not plot segments}

\item{root}{default TRUE to show root nodes, FALSE in case there are other root nodes that are not meaningful.}
}
\value{
Figure plotting stream network and observations
}
\description{
\code{plot_network} generates figure showing network nodes, directional arrows, and observations
}
