\name{VASTPlotUtils-package}
\alias{VASTPlotUtils-package}
\alias{VASTPlotUtils}
\docType{package}
\title{
\packageTitle{VASTPlotUtils}
}
\description{
\packageDescription{VASTPlotUtils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{VASTPlotUtils}
\packageIndices{VASTPlotUtils}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{VASTPlotUtils}

Maintainer: \packageMaintainer{VASTPlotUtils}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
