\name{VASTPlotUtils-package}
\alias{VASTPlotUtils-package}
\alias{VASTPlotUtils}
\docType{package}
\title{
\packageTitle{VASTPlotUtils}
}
\description{
\packageDescription{VASTPlotUtils}
}
\details{

\packageDESCRIPTION{VASTPlotUtils}

\author{
\packageAuthor{James Thorson, Merrill Rudd}

Maintainer: \packageMaintainer{Merrill Rudd}
}
}

